Add LEDs, Refactor AutoTurn, Add AutoDrive
Added LEDs to the Shooter subsystem. AutoTurn was refactored a bit; I added a new constructor that takes a target angle as a parameter, and I made it so we can easily tune the PID values via SmartDashboard. Added AutoDrive, which is just a renamed clone of AutoTurn for the time being.